+++ pkg/languagedetection/detector_linux_test.go
@@ -9,8 +9,11 @@ package languagedetection
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/config"
+	"github.com/DataDog/datadog-agent/pkg/process/net"
 	"github.com/stretchr/testify/assert"
 	"net/http"
+	"net/http/httptest"
+	"path"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -19,6 +22,22 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/languagedetection/languagemodels"
 	languagepb "github.com/DataDog/datadog-agent/pkg/proto/pbgo/languagedetection"
 )
+
+func startTestUnixServer(t *testing.T, handler http.Handler) string {
+	t.Helper()
+
+	socketPath := path.Join(t.TempDir(), "test.sock")
+	listener, err := net.NewListener(socketPath)
+	require.NoError(t, err)
+	t.Cleanup(listener.Stop)
+
+	srv := httptest.NewUnstartedServer(handler)
+	srv.Listener = listener.GetListener()
+	srv.Start()
+	t.Cleanup(srv.Close)
+
+	return socketPath
+}
 
 func TestBinaryAnalysisClient(t *testing.T) {
 	socketPath := startTestUnixServer(t, http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
